feat: introduce first super basic init command
This adds a new `init` command to initialize a Vibranium project in either the
current directory or a given optional path location:

```
$ vibranium init --path ../../path/to/project
```

If the directories in `path` don't exist, they'll be created recursively.
This is a super basic implementation and doesn't account for property interaction
for cases like trying to initialize a Vibranium project within a Vibranium
 project. If the needed directories already exist, Vibranium will just move
on.