feat: introduce reset command
This commit adds a new `reset` command to vibranium's CLI, which is very
handy to quickly wipe out all generated artifacts by Vibranium in the current
working directory:

```
$ vibranium reset
```

It also accepts an optional `path` argument so resetting project that are
located somewhere else on the disc is possible too:

```
$ vibranium reset --path path/to/project
```

This command will abort if the given location isn't a Vibranium project
(e.g. a dedicated .vibranium directory doesn't exist).