fix: handle error correctly when Vibranium config is corrupted
Prior to this commit it was theoretically possible to break Vibranium by
generating a project using

```
$ vibranium init --path some/path
```

And then invalidating its `vibranium.toml` by, e.g. removing any of the config
values.

The next time the project is attempted to be reset by Vibranium using

```
$ vibranium reset --path path/to/created/project
```

it'll try to read the existing config and panic as the config isn't in a valid shape.

With this commit a new project generation specific error type is introduced and
users will get a meaningful error message about why the configuration is invalid.