feat(vibranium::compiler): introduce rudimentary compilation support
With this commit Vibranium gets a new `compile` command that uses `solc` to compile a
project's Smart Contracts specified in the `vibranium.toml.

Similar to other commands, it comes with a `--path` option to specify a project path:

```
$ vibranium compile --path path/to/project
```

It's also possible to specify a different compiler binary than `solc` using the `--compiler` option:

```
$ vibranium compile --compiler vyper
```

However, this is will very likely fail as the actual compile functionality uses hard coded flags
from `solc`. This needs to be fixed using a strategy pattern in follow-up commits.

Lastly, it's possible to apply compiler specific options using the `--` syntax like this:

```
$ vibranium compile -- --abi --ast-json ...
```