fix(vibranium::compiler): ensure glob pattern is composed correctly
Prior to this commit there was a bug where the glob pattern for finding Smart Contract
source files wasn't composed correctly, when the given project path didn't have a trailing
slash.

For example, running:

```
$ vibranium compile --path ../some/path # no trailing slash
```

Caused the compiler algorithm to create "full" glob patterns like

```
../some/pathcontracts/*.sol
```

The reason for this is because the routine that aggregates the source files worked with
String types instead of PathBufs.

This commit changes the code to use PathBuf instead so missing slashes in path strings is
no longer a problem.