refactor(vibranium::compiler): introduce strategy pattern for compiler selection
This commit makes it more straight forward to add support for compilers. Vibranium's
`compile` command comes with a `--compiler` option that lets users specify which
compiler should be used for Smart Contract compilation:

```
$ vibranium compile --compiler solc
```

Prior to this commit, when supplying a compiler that isn't installed, Vibranium would
cancel its execution and tell the user that the compiler couldn't be found.

However, now with the strategy pattern in place, there's also the notion of "supported"
compilers. With this commit only the `solc` compiler is supported. When running the
compile command with a different compiler e.g.

```
$ vibranium compile --compiler vyper
```

Vibranium will tell the user that the given compiler isn't supported.