fix(vibranium::compiler): treat stderr as error output
Prior to this commit, when running

```
$ vibranium compile
```

And the underlying compiler would fail, Vibranium would still treat the
result as successful. This was because we didn't capture `Stdio` from
the underlying process.

With this commit we're introducing `compiler::error::CompilerError::Other(String)`,
capture `stdout` and `stderr from the underlying `Child` process and emit and
error with `stderr`.

At the same time we're moving some output messages around to reduce redundancy.

Fixes #2