feat: improve compiler option configuration with vibranium.toml
This enables users to specify compiler specific options in the vibranium.toml
configuration file. Possible options are `compiler.cmd` and `compiler.options`.

Example:

```
[compiler]
cmd = "solcjs"
options = ["--abi", "--optimize"]
```

This configuration overrules Vibranium's default configuration (e.g. Vibranium
uses `solc` by default). `compiler.options` are "merged" with default options
specified in the compiler.

Given a configuration like above, it's still possible to override these on a
command basis. E.g:

```
$ vibranium compile --compiler solc -- --standard-json
```

Will dismiss `compiler.cmd` and `compiler.options` in vibranium.toml.

It's also possible to only specify one of these options. For example having:

```
[compiler]
cmd = "solcjs"
```

And running

```
$ vibranium compile
```

Will just work, due to Vibranium's built-in strategy for `solcjs`.

The same way it's possible to just supply compiler options and set the compiler
program on the fly:

```
[compiler]
options = ["--abi"]
```
```
$ vibranium compile --compiler solcjs
```

This will still use the options defined in vibranium.toml. In fact, this could
be combined with a compiler that Vibranium doesn't have a compiler strategy for:

```
$ vibranium compile --compiler unsupported
```

Which would be the equivalent to running:

```
$ vibranium compile --compiler unsupported -- --abi
```