feat(vibranium::blockchain): introduce accounts commmand
This also introduces a new blockchain connector API which accepts an underlying
adapter implementation (e.g. web3).

APIs that use the blockchain connector rely on a blockchain connector configuration
which can be specified in a project's vibranium.toml file:

```
[blockchain.connector]
protocol = "rpc"
port = "4585"
host = "localhost"
```

This commit adds support for RPC and WebSocket support using a Web3Adapter.

The `accounts` command can be used as:

```
$ vibranium accounts
```

It's required to have a running blockchain process that Vibranium can connect
to otherwise this command will render an error message.

Closes #30 #31