refactor: simplify compiler by unifying the command pattern
Originally we've introduced a strategy pattern to let the compiler module
decide which program to run. However, since the crucial task will always
be the same (running a program in a child process and passing options to it),
there's actually no point in maintaining multiple strategies that practically
do the same thing.

This commit uses the same pattern used in the blockchain module and adds support
for running the command on windows machines as well.

Both modules can probably benefit from a unified abstraction as well as they are
very similar now.

Closes #28