feat: introduce deploy command
This commit introduces deployment capabilities, enabling users to configure
their Smart Contracts and having Vibranium deploy them accordingly.

There's a new, optional, `deployment` section in the vibranium.toml which
looks something like this:

```
[deployment]

  gas_price = 8
  gas_limit = 200000
  tx_confirmations = 0

  [[deployment.smart_contracts]]
    name = "SimpleStorage"
    args = [
      { value = "200", kind = "uint" },
    ]

  [[deployment.smart_contracts]]
    name = "AnotherStorage"
    args = [
      { value = "true", kind = "bool" },
      { value = "300", kind = "uint" },
    ]
```

There are a couple of things that can be configured here:

- `tx_confirmations` - Number of confirmations needed (defaults to 0 for development mode)
- `gas_price` - The general gas price used for all deployments (defaults to 5)
- `gas_limit` - The gas limit used for all deployments (defaults to 2000000)
- `smart_contracts[]` - Configurations for a DApp's Smart Contracts
  - `name` - The name of the Smart Contract
  - `args` - A list of constructor arguments for the Smart Contract

It's important to specify the `name` of the Smart Contracts so Vibranium
can identify then on the file system after compilation. If there are no artifacts,
Vibranium will simply not deploy anything.

Constructor arguments needed for Smart Contracts can be configured using
the `args` field. Vibranium will make sure that users aren't trying to provide
constructor arguments that are either invalid, unknown or simply mismatching in
type.

With a configuration given above, Smart Contracts can be deployed using

```
$ vibranium deploy
```

Or, as always, providing a path to a vibranium project:

```
$ vibranium deploy --path some/path/to/a/project
```

As Vibranium doesn't know the notion of environments yet, it'll assume
"development" as default and therefore deployment only works for blockchain
nodes on development mode.

Closes #46 36