feat(vibranium::deployment): introduce deployment tracking
With this, Vibranium starts tracking deployed Smart Contracts by hashing their
name, bytecode and arguments, and storing the data together with the first
block hash of a chain on disc.

Tracking is turned on by default and can be turned off using the new
`tracking_enabled` option in the deployment config:

```
[deployment]
  tracking_enabled: false # defaults to true
```

In case of a Smart Contract already being deployed in its current stage,
Vibranium will mark the Smart Contract in question as "skipped".

Closes #50