fix(vibranium::blockchain): provide Parity with dev account
This is needed to have parity process transactions. Parity requires the
options `--account` and `--password` which it will use to sign transactions
in dev mode.

`--password` is actually the path to a file that contains a list of passwords.
By default, vibranium will create a file `[project]/.vibranium/passwords/parity_dev`
with the password `dev_account` and the derivative address. That path is then used
unless `--password` is provided.

When users provide a custom path for `--password` they have to provide `--account`
as well.