feat(vibranium::deployment): introduce directives for referencing Smart Contracts
Users are now able to reference Smart Contracts as constructor arguments
of other Smart Contracts. The idea is that users can specify the address
of another Smart Contract using the `$CONTRACTNAME` syntax.

```
[[deployment.smart_contracts]]
  name = "SimpleStorage"
  args = [
    { value = "$AnotherStorage", kind = "address" },
  ]

[[deployment.smart_contracts]]
  name = "AnotherStorage"
  args = [
    { value = "true", kind = "bool" },
    { value = "200", kind = "uint" },
  ]
```

This enables vibranium to also determine the order in which Smart Contracts
need to be deployed. The example above will deploy `AnotherStorage`
before `SimpleStorage` and uses `AnotherStorage`'s address as a constructor
parameter for `SimpleStorage`.

Cyclic dependencies are not allowed and referencing Smart Contracts that
don't actually exist in the overall configuration will cause vibranium
to abort the deployment process.

Closes #57